Move metadata functions back into Set and Layer
That's where they are used...
